requirements.txt
asgiref==3.7.2
Django==3.2.6
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
